+++ client/src/components/About/presidentData.txt
